Move dataset ingest tests to separate file
